Rearrange validator test files
